[c_cdd/tests/test_function.h] Experiment with resolving infinite loop of `az_span_to_str` through `c_str_precondition_internal` on `az_span_to_str`; [c_cdd/{ll{.h,.c},cst.c}] Use `int32_t` over `size_t`; [.clang-format] Add back

+++ c_cdd/ll.c
@@ -90,7 +90,7 @@ struct az_span_elem **ll_append_span(struct az_span_elem **p, az_span span) {
   return ll_prepend_span(list_end(p), span);
 }
 
-struct az_span_elem **ll_push_span(size_t *ll_n, struct az_span_elem ***ll_root,
+struct az_span_elem **ll_push_span(int32_t *ll_n, struct az_span_elem ***ll_root,
                                    const az_span span) {
   if (az_span_ptr(span) != NULL && az_span_size(span) > 0) {
     (*ll_n)++;
